Added --purge-edges and --purge-vertices option
Filtered vertices and edges can be permanently removed from the graph
with --purge-vertices and --purge-edges, respectively. The edge or
vertex filter is automatically removed, afterwards. This is useful if
maximum speed is necessary, and saving and reloading the graph without
filtering is not desired.
    
(this commit also removes some trailing whitespaces)