Fix vertex and edge filtering, when only one is active
When only one of vertex or edge filtering was disabled, the allowed
range of the disabled filter was set to
]numeric_limits<double>::min(), numeric_limits<double>::max()[, and
the selected filtering property was the respective index. But
according to the STL documentation from GCC,
numeric_limits<>::min() returns:
    
   "The minimum finite value, or for floating types with
    denormalization, the minimum positive normalized value."
    
which is always positive for double (!), thus introducing a weird
regression, where the first vertex (index 0) is always filtered out if
only the edge filter is active, and vice-versa.