Further improvement of python interface
Vertices and edges can be accessed from the graph class, as such:

    import graph_tool
    g = graph_tool.Graph()
    for v in g.vertices():
        for e in v.out_edges():
           # do something...

Additionally, the --edit-{vertex|edge|graph}-property was ported to the
new interface, and is working again, as it used to.

The Vertex and Edge class no longer have the 'get_property' and
'set_property' method. They'll be replaced by a new method of accessing
property maps.